fixed loading script, implemented staging layer

+++ scripts/load_to_postgres.py
@@ -1,6 +1,7 @@
-from sqlalchemy import create_engine,text
+from sqlalchemy import create_engine,text,inspect
 from file_download import download_files
 from io import StringIO
+from pathlib import Path
 import pyarrow.parquet as pq
 import pandas as pd
 import os
@@ -13,40 +14,45 @@ engine = create_engine(
     f"{os.getenv('POSTGRES_DB')}"
                         )
 
+schema_name = 'raw'
 with engine.begin() as connection:
-    connection.execute(text('create schema if not exists raw'))
+    connection.execute(text(f'create schema if not exists {schema_name}'))
 
 files_dict = download_files()
 
 for name,path in files_dict.items():
 
-    parquet_file = pq.ParquetFile(path)
-    parquet_file_first_batch = next(parquet_file.iter_batches(10))
-    df_empty = parquet_file_first_batch.to_pandas(types_mapper=pd.ArrowDtype).head(0)
-    df_empty.to_sql(
-        name = name,
-        con = engine,
-        schema ='raw',
-        if_exists = 'replace',
-        index = False,
-    )
+    db_inspector = inspect(engine)
+    table_exists = db_inspector.has_table(table_name=name,schema =schema_name )
 
-    taxi_csv_path = f"/app/data/raw/{name}.csv"
+    if not table_exists:
+
+        parquet_file = pq.ParquetFile(path)
+        parquet_file_first_batch = next(parquet_file.iter_batches(10))
+        df_empty = parquet_file_first_batch.to_pandas(types_mapper=pd.ArrowDtype).head(0)
+        df_empty.to_sql(
+            name = name,
+            con = engine,
+            schema =schema_name,
+            if_exists = 'replace',
+            index = False,
+        )
+
+        taxi_csv_path = f"/app/data/raw/{name}.csv"
 
 
-    for batch_number,batch in enumerate(parquet_file.iter_batches(50000)):
-        batch_csv_buffer = StringIO()
-        batch_dataframe = batch.to_pandas(types_mapper=pd.ArrowDtype)
-        batch_dataframe.to_csv(batch_csv_buffer,
-                                mode = 'w' if batch_number==0 else 'a',
-                                header = True if batch_number==0 else False,
-                                index=False)
+        for batch_number,batch in enumerate(parquet_file.iter_batches(100000)):
+            batch_csv_buffer = StringIO()
+            batch_dataframe = batch.to_pandas(types_mapper=pd.ArrowDtype)
+            batch_dataframe.to_csv(batch_csv_buffer,
+                                    header = False,
+                                    index=False)
 
-    batch_csv_buffer.seek(0)
-    raw_connection = engine.raw_connection()
-    try:
-        with raw_connection.cursor() as psy_cursor:
-            psy_cursor.copy_expert(f"""COPY raw.{name} FROM STDIN with (format csv, header true)""",batch_csv_buffer)
-            raw_connection.commit()
-    finally:
-        raw_connection.close()
+            batch_csv_buffer.seek(0)
+            raw_connection = engine.raw_connection()
+            try:
+                with raw_connection.cursor() as psy_cursor:
+                    psy_cursor.copy_expert(f"""COPY {schema_name}.{name} FROM STDIN with (format csv)""",batch_csv_buffer)
+                    raw_connection.commit()
+            finally:
+                raw_connection.close()
